Proposed name change for fringe_tracing
